Chart view fix

+++ app/src/main/java/com/ferelin/stockprice/custom/utils/Marker.kt
@@ -1,3 +1,3 @@
 package com.ferelin.stockprice.custom.utils
 
-data class Marker(val currentPos: Point = Point(0f, 0f), val value: Int = 0)
+data class Marker(val currentPos: Point = Point(0f, 0f), val value: Float = 0F)
